Remove unnecessary RUBYOPT
This seems to be causing some issues with the new test, but it does not
seem needed. I don't recall why I added.

As long as our development binstubs load the rubygems version switcher
initially, the proper path to rubygems will be added to RUBYOPT and
preserved for all future subprocesses. I don't think further
manipulations are needed.